pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pywaybackup"]

[project]
name = "pywaybackup"
version = "4.0.0"
description = "Query and download archive.org as simple as possible."
authors = [
    { name = "bitdruid", email = "bitdruid@outlook.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pysqlite3-binary==0.5.4; sys_platform == 'linux'",
    "pysqlite-binary; sys_platform == 'win32'",
    "requests==2.32.3",
    "tqdm==4.67.1",
    "python-magic==0.4.27; sys_platform == 'linux'",
    "python-magic-bin==0.4.14; sys_platform == 'win32'",
]

[project.scripts]
waybackup = "pywaybackup.main:cli"

[project.urls]
homepage = "https://github.com/bitdruid/python-wayback-machine-downloader"